Moved ProductInfoHeadings into a more appropriate folder and also moved the rendering of that component to ProductCatalogDisplay
